i dont think we need this commit, BUT:
i did figure out that the latency issue was
on the receiver side, so the video pipeline here
seems to be tight enough to go ahead and move forward
to the webrtc stuff.

the reason i was getting latency was that vlc has a
large network cache default, so invoking it like this:

vlc --network-caching 90 ./test.sdp

and you achieve subsecond latency!

i finally found this command line argument whilst
searching around and finding this page:

https://eneus.github.io/blog/live-webcam-streaming-using-vlc-on-the-command-line/